+++ src/main/java/dev/isxander/controlify/compatibility/iris/IrisMixinPlugin.java
@@ -4,7 +4,7 @@ import dev.isxander.controlify.compatibility.CompatMixinPlugin;
 
 public class IrisMixinPlugin extends CompatMixinPlugin {
     @Override
-    public String getModId() {
+    protected String getModId() {
         return "iris";
     }
 }