+++ src/main/java/dev/isxander/controlify/compatibility/sodium/SodiumMixinPlugin.java
@@ -4,7 +4,7 @@ import dev.isxander.controlify.compatibility.CompatMixinPlugin;
 
 public class SodiumMixinPlugin extends CompatMixinPlugin {
     @Override
-    public String getModId() {
+    protected String getModId() {
         return "sodium";
     }
 }