+++ src/main/java/dev/isxander/controlify/compatibility/yacl/YACLMixinPlugin.java
@@ -4,7 +4,7 @@ import dev.isxander.controlify.compatibility.CompatMixinPlugin;
 
 public class YACLMixinPlugin extends CompatMixinPlugin {
     @Override
-    public String getModId() {
+    protected String getModId() {
         return "yet_another_config_lib_v3";
     }
 }